If a category attribute is linked with category then it cannot  be deleted

+++ WebAPI/VintageCars.Web/VintageCars.Service/Catalog/Handlers/DeleteCategoryAttributeHandler.cs
@@ -24,6 +24,9 @@ namespace VintageCars.Service.Catalog.Handlers
             var categoryAttribute = _extendedCategoryService.GetCategoryAttribute(request.Id)
                                     ?? throw new ValidationException(
                                         _localizationService.GetResource("DeleteCategoryAttribute.NotExists.Validation"));
+            if(_extendedCategoryService.HasAnyMappings(categoryAttribute.Id))
+                throw new ValidationException(_localizationService.GetResource("DeleteCategoryAttribute.IsLinkedWithCategory.Validation"));
+
             _extendedCategoryService.DeleteCategoryAttribute(categoryAttribute);
             return Unit.Task;
         }
